+++ common/rmi/src/main/java/is/codion/common/rmi/client/DefaultConnectionRequest.java
@@ -27,6 +27,7 @@ import java.time.ZoneId;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 
 import static java.util.Collections.emptyMap;
@@ -81,8 +82,8 @@ final class DefaultConnectionRequest implements ConnectionRequest, Serializable 
 	}
 
 	@Override
-	public Version clientVersion() {
-		return clientVersion;
+	public Optional<Version> clientVersion() {
+		return Optional.ofNullable(clientVersion);
 	}
 
 	@Override